ultima versao

+++ implementacao/TemplateVariation/src/main/java/br/edimarmanica/templatevariation/auto/similarity/SizeSimilarity.java
@@ -21,7 +21,7 @@ public class SizeSimilarity extends RuleSimilarity{
     
     @Override
     public double score() {
-        return candidateComplementaryRule.getUrlValues().size() / nrPages;
+        return (double)candidateComplementaryRule.getUrlValues().size() / nrPages;
     }
     
 }
